Initialize Jable_UI with GUI and cleaned environment

+++ delete.py
@@ -3,8 +3,9 @@ import os
 def deleteMp4(folderPath):
     files = os.listdir(folderPath)
     originFile = folderPath.split(os.path.sep)[-1] + '.mp4'
+    coverFile = folderPath.split(os.path.sep)[-1] + '.jpg'
     for file in files:
-        if file != originFile:
+        if file != originFile and file != coverFile:
             os.remove(os.path.join(folderPath, file))
